requirements.txt
httpx>=0.27
openai>=1.30
python-dotenv>=1.0
